backend/requirements.txt
# Core Backend
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1

# AI + Agents
crewai==0.55.2
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43

# LLM Providers
groq==0.9.0
openai==1.30.1
anthropic==0.28.0

# Search API
tavily-python

# RAG / Embeddings
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# Utilities
requests